requirements.txt
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.2.0
SQLAlchemy>=2.0.25
pydantic>=2.5.3
pytest>=7.4.3
python-dateutil>=2.8.2
requests>=2.31.0
